+++ lib/blocs/settings/settings_state.dart
@@ -28,6 +28,8 @@ class SettingsState extends Equatable {
   final bool groupTimers;
   final bool collapseDays;
   final bool autocompleteDescription;
+  final bool defaultFilterStartDateToMonday;
+  final bool oneTimerAtATime;
 
   SettingsState({
     @required this.exportGroupTimers,
@@ -42,6 +44,8 @@ class SettingsState extends Equatable {
     @required this.groupTimers,
     @required this.collapseDays,
     @required this.autocompleteDescription,
+    @required this.defaultFilterStartDateToMonday,
+    @required this.oneTimerAtATime,
   })  : assert(exportGroupTimers != null),
         assert(exportIncludeDate != null),
         assert(exportIncludeProject != null),
@@ -53,7 +57,9 @@ class SettingsState extends Equatable {
         assert(defaultProjectID != null),
         assert(groupTimers != null),
         assert(collapseDays != null),
-        assert(autocompleteDescription != null);
+        assert(autocompleteDescription != null),
+        assert(defaultFilterStartDateToMonday != null),
+        assert(oneTimerAtATime != null);
 
   static SettingsState initial() {
     return SettingsState(
@@ -69,10 +75,13 @@ class SettingsState extends Equatable {
       groupTimers: true,
       collapseDays: false,
       autocompleteDescription: true,
+      defaultFilterStartDateToMonday: false,
+      oneTimerAtATime: false,
     );
   }
 
-  SettingsState.clone(SettingsState project, {
+  SettingsState.clone(
+    SettingsState settings, {
     bool exportGroupTimers,
     bool exportIncludeDate,
     bool exportIncludeProject,
@@ -85,31 +94,31 @@ class SettingsState extends Equatable {
     bool groupTimers,
     bool collapseDays,
     bool autocompleteDescription,
-    })
-      : this(
-          exportGroupTimers: exportGroupTimers ?? project.exportGroupTimers,
-          exportIncludeDate:
-              exportIncludeDate ?? project.exportIncludeDate,
+    bool defaultFilterStartDateToMonday,
+    bool oneTimerAtATime,
+  }) : this(
+          exportGroupTimers: exportGroupTimers ?? settings.exportGroupTimers,
+          exportIncludeDate: exportIncludeDate ?? settings.exportIncludeDate,
           exportIncludeProject:
-              exportIncludeProject ?? project.exportIncludeProject,
+              exportIncludeProject ?? settings.exportIncludeProject,
           exportIncludeDescription:
-              exportIncludeDescription ?? project.exportIncludeDescription,
-          exportIncludeProjectDescription:
-              exportIncludeProjectDescription ?? project.exportIncludeProjectDescription,
+              exportIncludeDescription ?? settings.exportIncludeDescription,
+          exportIncludeProjectDescription: exportIncludeProjectDescription ??
+              settings.exportIncludeProjectDescription,
           exportIncludeStartTime:
-              exportIncludeStartTime ?? project.exportIncludeStartTime,
+              exportIncludeStartTime ?? settings.exportIncludeStartTime,
           exportIncludeEndTime:
-              exportIncludeEndTime ?? project.exportIncludeEndTime,
+              exportIncludeEndTime ?? settings.exportIncludeEndTime,
           exportIncludeDurationHours:
-              exportIncludeDurationHours ?? project.exportIncludeDurationHours,
-          defaultProjectID:
-              defaultProjectID ?? project.defaultProjectID,
-          groupTimers:
-              groupTimers ?? project.groupTimers,
-          collapseDays:
-              collapseDays ?? project.collapseDays,
+              exportIncludeDurationHours ?? settings.exportIncludeDurationHours,
+          defaultProjectID: defaultProjectID ?? settings.defaultProjectID,
+          groupTimers: groupTimers ?? settings.groupTimers,
+          collapseDays: collapseDays ?? settings.collapseDays,
           autocompleteDescription:
-              autocompleteDescription ?? project.autocompleteDescription,
+              autocompleteDescription ?? settings.autocompleteDescription,
+          defaultFilterStartDateToMonday: defaultFilterStartDateToMonday ??
+              settings.defaultFilterStartDateToMonday,
+          oneTimerAtATime: oneTimerAtATime ?? settings.oneTimerAtATime,
         );
 
   @override
@@ -126,5 +135,7 @@ class SettingsState extends Equatable {
         groupTimers,
         collapseDays,
         autocompleteDescription,
+        defaultFilterStartDateToMonday,
+        oneTimerAtATime,
       ];
 }