+++ lib/data_providers/settings/mock_settings_provider.dart
@@ -22,12 +22,13 @@ class MockSettingsProvider extends SettingsProvider {
   Locale locale;
 
   MockSettingsProvider() {
-    values = Map<String, dynamic>();
+    values = <String, dynamic>{};
     theme = ThemeType.auto;
   }
 
   @override
-  bool getBool(String key) => values.containsKey(key) ? values[key] as bool : true;
+  bool getBool(String key) =>
+      values.containsKey(key) ? values[key] as bool : true;
 
   @override
   void setBool(String key, bool value) => values[key] = value;