+++ CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-
+### Added
+* Added setting to automatically filter out timers older than the last Monday
+* Added setting to only allow a single active timer at once (when creating new ones)
 
 ## [1.5.0] - 2020-04-07
 ### Added